Refactor other pages; Change bottom navbar layout
