+++ uni/lib/view/common_widgets/pages_layouts/secondary/secondary.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:uni/view/common_widgets/pages_layouts/general/general.dart';
-import 'package:uni/view/common_widgets/pages_layouts/general/widgets/app_bar.dart';
 import 'package:uni/view/common_widgets/pages_layouts/general/widgets/bottom_navigation_bar.dart';
 import 'package:uni/view/common_widgets/pages_layouts/general/widgets/refresh_state.dart';
+import 'package:uni/view/common_widgets/pages_layouts/general/widgets/top_navigation_bar.dart';
 
 /// Page with a back button on top
 abstract class SecondaryPageViewState<T extends StatefulWidget>
